+++ packages/plugin-ext/src/main/browser/webview-views/webview-views.ts
@@ -25,13 +25,9 @@ import { WebviewWidget } from '../webview/webview';
 export interface WebviewView {
     title?: string;
     description?: string;
-    badge?: number | undefined;
-    badgeTooltip?: string | undefined;
     readonly webview: WebviewWidget;
     readonly onDidChangeVisibility: Event<boolean>;
     readonly onDidDispose: Event<void>;
-    readonly onDidChangeBadge: Event<void>;
-    readonly onDidChangeBadgeTooltip: Event<void>;
 
     dispose(): void;
     show(preserveFocus: boolean): void;